[FEATURE] Get-Request für die Averagenote + TrainingModuleWithAverageDto + weighting von float auf int geändert

+++ Assessify.Backend/src/main/java/de/assessify/app/assessifyapi/api/dtos/request/UpdateGradeDto.java
@@ -2,4 +2,4 @@ package de.assessify.app.assessifyapi.api.dtos.request;
 
 import java.util.Date;
 
-public record UpdateGradeDto (String gradeName, Float value, Float gradeWeighting, Date date){}
+public record UpdateGradeDto (String gradeName, float value, int gradeWeighting, Date date){}
